// app/src/types/theme.ts
export type ThemeFonts = {
    regular: string;
    light: string;
    semiBold: string;
}

export type Theme = {
    backgroundColor: string;
    fonts: ThemeFonts;
    inputTextColor: string;
    inputBackgroundColor: string;
    borderColor: string;
    placeholderColor: string;
    errorColor: string;
    backgroundErrorColor: string;
    disabledButtonColor: string;
}